+++ app/tasks/schemas/task_event_metric_schema.py
@@ -1,10 +1,12 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class TaskEventMetricCreationSchema(BaseModel):
     task_metric_id: int
     task_event_id: int
-    value: int
+    value: Decimal
 
 
 class TaskEventMetricSchema(TaskEventMetricCreationSchema):
